fix: refactor actions code for all user actions to be in a single file. change all necessarry imports referring to the actions to the newly formated file
