+++ client/src/components/account/Login.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { PropTypes } from 'prop-types';
 import HomePage from '../home/HomePage';
 import { connect } from 'react-redux';
-import { LoginAction } from '../../actions/LoginAction';
+import { LoginAction } from '../../actions/authActions';
 import { addFlashMessage } from '../../actions/flashMessages';
 
 /**
