+++ client/src/components/account/SIgnUpForm.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { PropTypes } from 'prop-types'; 
 import { Link } from 'react-router-dom';
-import { SignUpAction } from "../../actions/SignUpAction";
+import { SignUpAction } from "../../actions/authActions";
 import classnames from 'classnames';
 
 class SignUpForm extends React.Component {
@@ -20,7 +20,7 @@ class SignUpForm extends React.Component {
   };
   
 
-  
+
   onChange(e) {
     this.setState({ [e.target.name]: e.target.value });
   }
