+++ client/src/helpers/setAuth.js
